Refactor scripts
