+++ packages/www/_scripts/wipe-desc.ts
@@ -1,14 +1,13 @@
-import { prepareResources, Resource, writeMarkdown } from "./util"
+import { writeMarkdown } from "./util"
+import { getResources, Resource } from "./get-resources"
 
-const directories = ["db", "languages", "os", "tools"]
-const resources: Resource[] = prepareResources(directories)
-const nonIndexResources = resources.filter(r => !r.isIndex)
+const resources: Resource[] = getResources()
+const resourcesWithPrompts = resources.filter(r => r.prompt && !r.isIndex)
 
 async function wipeDesc() {
-  console.log(`Wiping desc...`)
-  for (const resource of nonIndexResources) {
+  for (const resource of resourcesWithPrompts) {
     resource.source.data.desc = null
-    writeMarkdown(resource.path, resource.source.content, resource.source.data)
+    writeMarkdown(resource.source.path, resource.source.content, resource.source.data)
   }
   console.log()
 }
